RAC-5474: To address issue when publish vagrant
(1) to resolve ERROR: [Errno 8] _ssl.c:510: EOF occurred in violation of
protocol
(2) to depress Warning:
python2.7/site-packages/urllib3/util/ssl_.py:335: SNIMissingWarning: An
HTTPS request has been made, but the SNI (Subject Name Indication)
extension to TLS is not available on this platform. This may cause the
server to present an incorrect TLS certificate, which can cause
validation failures. You can upgrade to a newer version of Python to
solve this. For more information, see
https://urllib3.readthedocs.io/en/latest/advanced-usage.html#ssl-warnings
  SNIMissingWarning